fix(go): colocate dual-stack transport helpers in exchange.go
Move newDualStackDialer/newDualStackTransport out of the standalone
exchange_transport.go file into exchange.go next to the HTTP client
setup. No behavior change.
